fix: add support for shorthand in repository queries
This reverts b07e335 and is necessary so that we still
destructure shorthand syntax if we pass a filter_group
directly into a Repository instance.

For example, while handling a request and you need to
query for and interact with a different resource.

- e.g. handling a POST to /comments, and you want to increment
a 'comment_count' on a different resource.